Refractor existing event types and event::poll, update test
Events are now individual *separate* types. The conversion between
wsdl2 typed structures and SDL structures is through and
incredibly dangerous reinterpert_cast<T*>(&e.t).

Therefore we need to test the shit out of this to make sure that
it never crashes.

To get an event the event::poll() function returns an optional
variant of the various events that can be checked in many ways.
Here's an example (requires <type_traits>, <variant>):

    auto event = event::poll();
    if (event) {
        std::visit([&](auto& arg) {
            using T = std::decay_t<decltype(arg)>;

            if constexpr (std::is_same_v<T, event::quit>) {
                // exit
            }
        }, *event);
    }

See std::visit() on cppreference, for more details.